feat(posts): add search functionality and role-based authorization

+++ src/useCases/delete-posts.ts
@@ -1,14 +1,20 @@
 import { IPostsRepository } from '@/repositories/posts.repository.interface'
 import { ResourceNotFoundError } from './errors/resource-not-found-error'
+import { UserRole } from '@/entities/enums/user-role'
+import { UnauthorizedError } from './errors/UnauthorizedError'
 
 export class DeletePostsUseCase {
   constructor(private postsRepository: IPostsRepository) {}
 
-  async execute(id: number): Promise<void> {
+  async execute(id: number, role: string): Promise<void> {
     const post = await this.postsRepository.findById(id)
 
     if (!post) {
       throw new ResourceNotFoundError()
+    }
+
+    if (role !== UserRole.PROFESSOR) {
+      throw new UnauthorizedError()
     }
 
     return this.postsRepository.delete(id)
